+++ ecc/app/graphrag/community_summarizer.py
@@ -1,16 +1,16 @@
 # Copyright (c) 2024-2026 TigerGraph, Inc.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
+# This program may be redistributed and/or modified under the terms of the GNU
+# Affero General Public License as published by the Free Software Foundation,
+# either version 3 of the License, or (at your option) any later version.
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+# details.
 #
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 import re
 import logging
@@ -19,6 +19,7 @@ from langchain_core.prompts import PromptTemplate
 from langchain_core.output_parsers import PydanticOutputParser
 
 from common.llm_services import LLM_Model
+from common.llm_services.base_llm import classify_llm_error
 from common.py_schemas import CommunitySummary
 
 logger = logging.getLogger(__name__)
@@ -55,5 +56,10 @@ class CommunitySummarizer:
                 caller_name="community_summarize",
             )
         except Exception as e:
-            return {"error": True, "summary": "", "message": str(e)}
+            return {
+                "error": True,
+                "summary": "",
+                "message": str(e),
+                "category": classify_llm_error(e),
+            }
         return {"error": False, "summary": summary.summary}